# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go-Playground - Random Go/Gin-Boilerplate Playground",
        "title": "Go-Playground",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/points/{customer_id}/{program_id}/balance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "UserIdAuth": []
                    }
                ],
                "description": "Get current points balance for a customer in a program",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "points"
                ],
                "summary": "Get points balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customer_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Program ID",
                        "name": "program_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/points/{customer_id}/{program_id}/earn": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "UserIdAuth": []
                    }
                ],
                "description": "Add points to a customer's balance in a program",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "points"
                ],
                "summary": "Earn points",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customer_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Program ID",
                        "name": "program_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Points to earn",
                        "name": "points",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.EarnPointsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/points/{customer_id}/{program_id}/ledger": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "UserIdAuth": []
                    }
                ],
                "description": "Get points ledger entries for a customer in a program",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "points"
                ],
                "summary": "Get points ledger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customer_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Program ID",
                        "name": "program_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.PointsLedger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/points/{customer_id}/{program_id}/redeem": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "UserIdAuth": []
                    }
                ],
                "description": "Redeem points from a customer's balance in a program",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "points"
                ],
                "summary": "Redeem points",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customer_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Program ID",
                        "name": "program_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Points to redeem",
                        "name": "points",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RedeemPointsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.EarnPointsRequest": {
            "type": "object",
            "required": [
                "points"
            ],
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "program_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.PointsLedger": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "ledger_id": {
                    "type": "string"
                },
                "merchant_customers_id": {
                    "type": "string"
                },
                "points_balance": {
                    "type": "integer"
                },
                "points_earned": {
                    "type": "integer"
                },
                "points_redeemed": {
                    "type": "integer"
                },
                "program_id": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.RedeemPointsRequest": {
            "type": "object",
            "required": [
                "points"
            ],
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "program_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "UserIdAuth": {
            "description": "User ID for authentication",
            "type": "apiKey",
            "name": "X-User-Id",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        },
        {
            "UserIdAuth": []
        }
    ]
}